% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_country_name.R
\name{alternative_country_name}
\alias{alternative_country_name}
\title{alternative_country_name}
\usage{
alternative_country_name(.x)
}
\arguments{
\item{.x}{a character vector of length one.}
}
\value{
a data frame
}
\description{
Returns alternative names for a common English country name.
}
\details{
Souce for common english country names and alternative country names: \url{"https://en.wikipedia.org/wiki/List_of_alternative_country_names"}.
}
\examples{

alternative_country_name("United states")

}
