% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_country_name.R
\name{simple_country_name}
\alias{simple_country_name}
\title{simple_country_name}
\usage{
simple_country_name(.x)
}
\arguments{
\item{.x}{a character vector of country names}
}
\value{
a data frame
}
\description{
Translates a vector of country names to common English equivalent. If no match is found the element is returned.
}
\details{
Source for common english country names is R package `countrycode` and alternative country names from `countrycode` and Wikipedia: \url{"https://en.wikipedia.org/wiki/List_of_alternative_country_names"}.
}
\examples{

x <- c("Czech", "Great Britain")
simple_country_name(x)

}
